Move about page to subfolder
